refactor pm quality policy action route
