+++ src/app/router_registry.py
@@ -95,6 +95,9 @@ from app.routers.dpm_command_center_pm_quality_fairness_lookup import (
 from app.routers.dpm_command_center_pm_quality_policies import (
     router as dpm_command_center_pm_quality_policies_router,
 )
+from app.routers.dpm_command_center_pm_quality_policy_actions import (
+    router as dpm_command_center_pm_quality_policy_actions_router,
+)
 from app.routers.dpm_command_center_pm_quality_review_action_lookup import (
     router as dpm_command_center_pm_quality_review_action_lookup_router,
 )
@@ -315,6 +318,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_command_center_pm_quality_fairness_router)
     app.include_router(dpm_command_center_pm_quality_fairness_lookup_router)
     app.include_router(dpm_command_center_pm_quality_policies_router)
+    app.include_router(dpm_command_center_pm_quality_policy_actions_router)
     app.include_router(dpm_command_center_pm_quality_review_actions_router)
     app.include_router(dpm_command_center_pm_quality_review_action_lookup_router)
     app.include_router(dpm_command_center_pm_quality_score_run_lookup_router)
